docs: integrate changelogs
