+++ apps/docs/src/development/packages/changelogs/[name].paths.ts
@@ -1,0 +1,10 @@
+import { getOnyxNpmPackages } from "../../../.vitepress/utils";
+
+// https://vitepress.dev/guide/routing#dynamically-generating-paths
+export default {
+  paths() {
+    return getOnyxNpmPackages().map((name) => {
+      return { params: { name } };
+    });
+  },
+};
